feat: store 폴더 생성 및 reducer 구현
